+++ src/main/kotlin/com/atlassian/performance/tools/virtualusers/LoadSegment.kt
@@ -1,5 +1,6 @@
 package com.atlassian.performance.tools.virtualusers
 
+import com.atlassian.performance.tools.jiraactions.api.memories.User
 import com.atlassian.performance.tools.virtualusers.api.browsers.CloseableRemoteWebDriver
 import org.apache.logging.log4j.LogManager
 import java.io.BufferedWriter
@@ -12,6 +13,7 @@ import java.util.concurrent.atomic.AtomicBoolean
 internal class LoadSegment(
     val driver: CloseableRemoteWebDriver,
     val output: BufferedWriter,
+    val user: User,
     val done: AtomicBoolean,
     val id: UUID,
     val index: Int