+++ src/main/kotlin/com/atlassian/performance/tools/virtualusers/api/users/UserGenerator.kt
@@ -4,6 +4,11 @@ import com.atlassian.performance.tools.jiraactions.api.memories.User
 import com.atlassian.performance.tools.virtualusers.api.VirtualUserOptions
 import net.jcip.annotations.ThreadSafe
 
+/**
+ * Supplies a user. The user can be created dynamically or it can find an existing user.
+ * Expect an instance to be called concurrently.
+ * If the implementation needs to be called sequentially, implement synchronization internally.
+ */
 @ThreadSafe
 interface UserGenerator {